Fixed proposal status option colors
